requirements.txt
Django==3.2.9
djangorestframework==3.13.1
pycodestyle==2.8.0
psycopg2==2.9.3
django-cors-headers==3.11.0